athena/tests/integration_tests/eunomia/pyproject.toml
[tool.poetry]
name = "eval"
version = "0.1.0"
description = ""
authors = ["leon <leon.wehrhahn@tum.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.11"
pyyaml = "6.0.2"
matplotlib = "3.9.3"
tabulate = "0.9.0"
pandas = "2.2.3"
requests = "2.32.3"
rich = "13.9.4"
questionary = "2.0.1"

[tool.poetry.group.dev.dependencies]
pydantic = "1.10.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
eval = "eval.main:main"